Extract common UI code to `VariationsBulkUpdateBaseFragment`
Introduced `VariationsBulkUpdateBaseFragment` - a base class for fragments related to variations bulk update attributes. It implements common UI parts like "Done" button menu item, progress dialog showing, confirmation snackbar, and navigation events handling.
